added listener file logic for modularity.
listener is used in order to seperate the coupling of like and post.
changed the api route too since it was not following the RESTApi concept,
An action should be associated to the resource it affects.
like is action and like is resource(child of post)
TASK DONE : like logic + likeCount in post.